requirements.txt
PyYaml
torch
numpy